+++ jena-fuseki-kafka-module/src/main/java/org/apache/jena/fuseki/kafka/FKS.java
@@ -16,6 +16,8 @@
 
 package org.apache.jena.fuseki.kafka;
 
+import io.opentelemetry.api.metrics.Meter;
+import io.telicent.smart.cache.observability.TelicentMetrics;
 import io.telicent.smart.cache.payloads.RdfPayload;
 import io.telicent.smart.cache.projectors.Sink;
 import io.telicent.smart.cache.projectors.driver.ProjectorDriver;
@@ -37,6 +39,7 @@ import org.apache.jena.fuseki.server.DataAccessPointRegistry;
 import org.apache.jena.fuseki.server.DataService;
 import org.apache.jena.kafka.JenaKafkaException;
 import org.apache.jena.kafka.KConnectorDesc;
+import org.apache.jena.kafka.SysJenaKafka;
 import org.apache.jena.kafka.common.FusekiOffsetStore;
 import org.apache.jena.kafka.common.FusekiProjector;
 import org.apache.jena.kafka.common.FusekiSink;
@@ -51,6 +54,7 @@ import org.apache.kafka.common.utils.Bytes;
 import java.time.Duration;
 import java.util.*;
 import java.util.concurrent.*;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 
 import static org.apache.jena.kafka.FusekiKafka.LOG;
@@ -59,6 +63,7 @@ import static org.apache.jena.kafka.FusekiKafka.LOG;
  * Functions for Fuseki-Kafka server setup.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
+@SuppressWarnings("resource")
 public class FKS {
 
     /**
@@ -188,7 +193,7 @@ public class FKS {
             Thread.currentThread().interrupt();
             throw new JenaKafkaException(
                     String.format("[%s] Interrupted while performing strict startup topic checks", topicNames), e);
-        } catch (JenaKafkaException e){
+        } catch (JenaKafkaException e) {
             throw e;
         } catch (RuntimeException e) {
             throw new JenaKafkaException(
@@ -254,12 +259,84 @@ public class FKS {
         return Collections.unmodifiableList(topics);
     }
 
+    /**
+     * A record that associates a name with a future to aid debugging
+     *
+     * @param name   Name
+     * @param future Future
+     */
+    public static record NamedFuture(String name, Future<?> future) {
+
+    }
+
     // All the static fields for managing the Kafka polling threads
     private static final Map<String, List<ProjectorDriver<Bytes, RdfPayload, Event<Bytes, RdfPayload>>>> DRIVERS =
             new ConcurrentHashMap<>();
-    private static final List<Future<?>> ACTIVE_DRIVERS = new CopyOnWriteArrayList<>();
+    private static final List<NamedFuture> ACTIVE_DRIVERS = new CopyOnWriteArrayList<>();
     private static PollThreadMonitor MONITOR;
     private static ExecutorService EXECUTOR = threadExecutor();
+    private static final AtomicInteger LAUNCHED = new AtomicInteger(0);
+    private static final AtomicInteger FAILED = new AtomicInteger(0);
+
+    /**
+     * Prefix to the metric names used to report the {@link #launched()}, {@link #failed()} and {@link #running()}
+     * metrics to OpenTelemetry
+     */
+    public static final String OTEL_POLL_THREADS_METRIC_PREFIX = "fuseki.kafka.poll.threads.";
+
+    static {
+        Meter meter = TelicentMetrics.getMeter("fuseki-kafka", SysJenaKafka.VERSION);
+        meter.gaugeBuilder(OTEL_POLL_THREADS_METRIC_PREFIX + "launched")
+             .ofLongs()
+             .setDescription("Number of Kafka polling threads that have been launched")
+             .buildWithCallback(measurement -> measurement.record(FKS.launched()));
+        meter.gaugeBuilder(OTEL_POLL_THREADS_METRIC_PREFIX + "failed")
+             .ofLongs()
+             .setDescription("Number of Kafka polling threads that have failed")
+             .buildWithCallback(measurement -> measurement.record(FKS.failed()));
+        meter.gaugeBuilder(OTEL_POLL_THREADS_METRIC_PREFIX + "running")
+             .ofLongs()
+             .setDescription("Number of Kafka polling threads that are actively running")
+             .buildWithCallback(measurement -> measurement.record(FKS.running()));
+    }
+
+    /**
+     * Gets how many, if any, polling threads have been launched
+     * <p>
+     * By comparing this and {@link #failed()} a caller can determine both whether any Fuseki Kafka processing is
+     * running and whether it is healthy.
+     * </p>
+     *
+     * @return Number of launched polling threads
+     */
+    public static int launched() {
+        return LAUNCHED.get();
+    }
+
+    /**
+     * Gets how many, if any, polling threads have failed
+     *
+     * @return Number of failed polling threads
+     */
+    public static int failed() {
+        return FAILED.get();
+    }
+
+    /**
+     * Gets how many actively running polling threads exist
+     * <p>
+     * Note that this may not return the same value as calculating it from <code>launched() - failed()</code> as this
+     * only reports actively running threads and a failure may not have been spotted by the monitoring thread and used
+     * to update those counts yet.  Also, if the service was in the process of shutting down then the threads may have
+     * been cancelled and/or naturally completed and that is not considered a failure so a naive calculation would
+     * return a higher number of active threads than there actually are.
+     * </p>
+     *
+     * @return Actively running threads
+     */
+    public static int running() {
+        return ACTIVE_DRIVERS.stream().filter(f -> !f.future().isDone()).mapToInt(i -> 1).sum();
+    }
 
     private static ExecutorService threadExecutor() {
         ExecutorService executor = Executors.newVirtualThreadPerTaskExecutor();
@@ -273,7 +350,12 @@ public class FKS {
 
 
     /**
-     * The background threads
+     * Resets all polling threads and associated thread execution machinery, explicitly cancelling threads wherever
+     * possible.
+     * <p>
+     * Intended primarily for use within unit and integration tests to allow resetting the Fuseki Kafka machinery
+     * between tests.
+     * </p>
      */
     static void resetPollThreads() {
         // Explicitly cancel the projector drivers
@@ -296,13 +378,16 @@ public class FKS {
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
         }
+        LAUNCHED.set(0);
+        FAILED.set(0);
         EXECUTOR = threadExecutor();
     }
 
-    private static void startTopicPoll(KConnectorDesc connector, KafkaRdfPayloadSource<Bytes> source,
-                                       DatasetGraph destination,
-                                       Function<DatasetGraph, Sink<Event<Bytes, RdfPayload>>> sinkBuilder) {
+    static void startTopicPoll(KConnectorDesc connector, KafkaRdfPayloadSource<Bytes> source,
+                               DatasetGraph destination,
+                               Function<DatasetGraph, Sink<Event<Bytes, RdfPayload>>> sinkBuilder) {
 
+        String topicNamesLabel = "[" + StringUtils.join(connector.getTopics(), ", ") + "]";
         //@formatter:off
         Sink<Event<Bytes, RdfPayload>> dlq = null;
         if (StringUtils.isNotBlank(connector.getDlqTopic())) {
@@ -315,6 +400,8 @@ public class FKS {
                             // NB - We want any failures in the DLQ to surface immediately
                             .noAsync()
                             .build();
+        } else {
+            LOG.warn("{} Connector does not configure a DLQ topic, if any malformed/unprocessable events are encountered this will cause the polling thread to fail and stop processing further events", topicNamesLabel);
         }
         ProjectorDriver<Bytes, RdfPayload, Event<Bytes, RdfPayload>> driver =
                 ProjectorDriver.<Bytes, RdfPayload, Event<Bytes, RdfPayload>>create()
@@ -324,7 +411,8 @@ public class FKS {
                                .source(source)
                                // Label each driver with the topics it pertains to, this makes the logs easier to read
                                // if there are multiple connectors defined
-                               .logLabel("[" + StringUtils.join(connector.getTopics(), ", ") + "]")
+                               .logLabel(topicNamesLabel)
+                               .threadName("Fuseki-Kafka Connector " + topicNamesLabel)
                                .projector(FusekiProjector.builder()
                                                          .source(source)
                                                          .dataset(destination)
@@ -339,6 +427,7 @@ public class FKS {
 
         // Submit for execution, and register for cancellation
         Future<?> future = EXECUTOR.submit(driver);
+        LAUNCHED.incrementAndGet();
         DRIVERS.computeIfAbsent(connector.getDatasetName(), x -> new CopyOnWriteArrayList<>()).add(driver);
 
         // Wait briefly for the projector driver thread to spin up
@@ -348,10 +437,12 @@ public class FKS {
             Thread.currentThread().interrupt();
             future.cancel(true);
             DRIVERS.getOrDefault(connector.getDatasetName(), Collections.emptyList()).remove(driver);
+            FAILED.incrementAndGet();
             throw new JenaKafkaException("Interrupted while waiting for connector to start up", e);
         } catch (ExecutionException e) {
             // If the projector driver fails in the startup phase we should bail out immediately
             DRIVERS.getOrDefault(connector.getDatasetName(), Collections.emptyList()).remove(driver);
+            FAILED.incrementAndGet();
             throw new JenaKafkaException("Connector failed to start up", e);
         } catch (TimeoutException e) {
             // Ignore, we can safely assume the projector driver thread started up cleanly
@@ -359,7 +450,7 @@ public class FKS {
 
         // Retain a reference to the future so our monitor thread can periodically check in on the drivers to see if
         // any have exited unexpectedly
-        ACTIVE_DRIVERS.add(future);
+        ACTIVE_DRIVERS.add(new NamedFuture(topicNamesLabel, future));
     }
 
     /**
@@ -508,14 +599,14 @@ public class FKS {
         private boolean shouldRun = true;
         private final Semaphore waitLock = new Semaphore(0);
         private final long checkInterval;
-        private final Collection<Future<?>> active;
+        private final Collection<NamedFuture> active;
 
         /**
          * Creates a new monitoring thread
          *
          * @param active Active thread futures to monitor
          */
-        public PollThreadMonitor(Collection<Future<?>> active, long checkInterval) {
+        public PollThreadMonitor(Collection<NamedFuture> active, long checkInterval) {
             this.active = Objects.requireNonNull(active);
             if (checkInterval < 0) {
                 throw new IllegalArgumentException("Check interval cannot be zero/negative");
@@ -535,12 +626,13 @@ public class FKS {
         }
 
         private void monitorActiveDrivers() {
-            for (Future<?> future : active) {
+            for (NamedFuture future : active) {
                 monitorDriver(future);
             }
         }
 
-        private void monitorDriver(Future<?> future) {
+        private void monitorDriver(NamedFuture named) {
+            Future<?> future = named.future();
             if (future.isCancelled()) {
                 // Ignore threads that have been explicitly cancelled
                 return;
@@ -552,7 +644,7 @@ public class FKS {
 
                 // If we successfully get() its result then it has exited normally
                 // BUT in most cases the thread should only exit if it either fails, or we're being shutdown
-                LOG.info("Polling thread exited normally");
+                LOG.info("{} Polling thread exited normally", named.name());
             } catch (ExecutionException e) {
                 // This means something fatal has happened on the polling thread to cause it to exit with an
                 // exception, log an error for this
@@ -560,7 +652,8 @@ public class FKS {
                 // an error on a ProjectorDriver thread and the ProjectorDriver will only log the basic error
                 // message.  If we don't log the stack trace we have zero visibility into what the system was
                 // doing when the error occurred making it difficult to debug.
-                LOG.error("Polling thread failed: ", e.getCause());
+                LOG.error("{} Polling thread failed: ", named.name(), e.getCause());
+                FAILED.incrementAndGet();
             } catch (InterruptedException e) {
                 stopAfterInterrupt();
             } catch (CancellationException | TimeoutException e) {
@@ -572,7 +665,7 @@ public class FKS {
         private void pruneCompletedDrivers() {
             // Remove any cancelled/failed/completed drivers from subsequent monitoring otherwise we'd spam the logs
             // with the error repeatedly
-            active.removeIf(f -> f.isCancelled() || f.isDone());
+            active.removeIf(f -> f.future().isCancelled() || f.future().isDone());
         }
 
         @SuppressWarnings("java:S899")